validationcaches: Use size_t for sizes
Summary:
...also move the 0-clamping logic to ApplyArgsManOptions, where it belongs.

This concludes backport of [[https://github.com/bitcoin/bitcoin/pull/25527 | core#25527]]
https://github.com/bitcoin/bitcoin/pull/25527/commits/0f3a2532c38074dd9789d1c4c667db6ca46ff0ab

Depends on D15378

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15740

+++ src/kernel/validation_cache_sizes.h
@@ -8,13 +8,13 @@
 #include <script/scriptcache.h>
 #include <script/sigcache.h>
 
-#include <cstdint>
+#include <cstddef>
 #include <limits>
 
 namespace kernel {
 struct ValidationCacheSizes {
-    int64_t signature_cache_bytes{DEFAULT_MAX_SIG_CACHE_BYTES};
-    int64_t script_execution_cache_bytes{DEFAULT_MAX_SCRIPT_CACHE_BYTES};
+    size_t signature_cache_bytes{DEFAULT_MAX_SIG_CACHE_BYTES};
+    size_t script_execution_cache_bytes{DEFAULT_MAX_SCRIPT_CACHE_BYTES};
 };
 } // namespace kernel
